Moved code to __main__.py
