+++ ingress-{{cookiecutter.repo_name}}/run.bat
@@ -3,5 +3,5 @@
 CALL E:\ingress-{{cookiecutter.repo_name}}\venv\Scripts\activate.bat
 PUSHD E:
 CD E:\ingress-{{cookiecutter.repo_name}}
-CALL python -m ingress_{{cookiecutter.module_name}}.adapter
+CALL python -m ingress_{{cookiecutter.module_name}}
 CALL E:\ingress-{{cookiecutter.repo_name}}\venv\Scripts\deactivate.bat
